# openapi.yml
openapi: 3.0.0
info:
  title: Chess Game API
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /games:
    get:
      summary: Get a list of all games
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: The unique ID of the game
                    state:
                      type: string
                      description: The FEN representation of the game's state
                    player1:
                      type: string
                      description: The name of player 1
                    player2:
                      type: string
                      description: The name of player 2
    post:
      summary: Create a new game
      requestBody:
        description: The details of the new game
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  description: The FEN representation of the game's initial state
                player1:
                  type: string
                  description: The name of player 1
                player2:
                  type: string
                  description: The name of player 2
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The unique ID of the new game
  /games/{game_id}:
    get:
      summary: Get the details of a game by ID
      parameters:
        - name: game_id
          in: path
          description: The unique ID of the game
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The unique ID of the game
                  state:
                    type: string
                    description: The FEN representation of the game's state
                  player1:
                    type: string
                    description: The name of player 1
                  player2:
                    type: string
                    description: The name of player 2
        "404":
          description: Game not found
    put:
      summary: Update the state of a game by ID
      parameters:
        - name: game_id
          in: path
          description: The unique ID of the game
          required: true
          schema:
            type: string
      requestBody:
        description: The updated details of the game
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  description: The FEN representation of the game's new state
      responses:
        "204":
          description: No Content
        "404":
          description: Game not found
    delete:
      summary: Delete a game by ID
      parameters:
        - name: game_id
          in: path
          description: The unique ID of the game
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Game not found